VACMS-18444 FL: Move action types into the actions folder
