+++ src/applications/facility-locator/actions/actionTypes.js
