+++ src/applications/facility-locator/actions/locations/getProviderSpecialties.js
@@ -3,7 +3,7 @@ import {
   FETCH_SPECIALTIES,
   FETCH_SPECIALTIES_FAILED,
   FETCH_SPECIALTIES_DONE,
-} from '../../utils/actionTypes';
+} from '../actionTypes';
 
 /**
  * Preloads all specialties available from CC Providers
