+++ src/applications/facility-locator/actions/mapbox/clearGeocodeError.js
@@ -1,4 +1,4 @@
-import { GEOCODE_CLEAR_ERROR } from '../../utils/actionTypes';
+import { GEOCODE_CLEAR_ERROR } from '../actionTypes';
 
 export const clearGeocodeError = () => async dispatch => {
   dispatch({ type: GEOCODE_CLEAR_ERROR });
