+++ src/applications/facility-locator/actions/mapbox/mapMoved.js
@@ -1,4 +1,4 @@
-import { MAP_MOVED } from '../../utils/actionTypes';
+import { MAP_MOVED } from '../actionTypes';
 
 export const mapMoved = currentRadius => ({
   type: MAP_MOVED,
