+++ src/applications/facility-locator/actions/search/clearSearchResults.js
@@ -1,4 +1,4 @@
-import { CLEAR_SEARCH_RESULTS } from '../../utils/actionTypes';
+import { CLEAR_SEARCH_RESULTS } from '../actionTypes';
 
 export const clearSearchResults = () => ({
   type: CLEAR_SEARCH_RESULTS,
