+++ src/applications/facility-locator/actions/search/clearSearchText.js
@@ -1,4 +1,4 @@
-import { CLEAR_SEARCH_TEXT } from '../../utils/actionTypes';
+import { CLEAR_SEARCH_TEXT } from '../actionTypes';
 
 export const clearSearchText = () => async dispatch => {
   dispatch({ type: CLEAR_SEARCH_TEXT });
